Profile, Disabled user validation and templates

See merge request rudypalacios/base-symfony-project!1

+++ src/Controller/UserController.php
@@ -100,9 +100,13 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user, UserRepository $userRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
-            $userRepository->remove($user, true);
+        $loggedUser = $this->getUser();
+        if ($user->getId() !== $loggedUser->getId()) {
+            if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+                $userRepository->remove($user, true);
+            }
         }
+
 
         return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
     }
